Python Website Development/FunOlympic/requirements.txt
asgiref==3.7.2
click==8.1.3
colorama==0.4.6
cssbeautifier==1.14.7
Django==4.2.1
django-admin-argon-dashboard==1.0.16
django-recaptcha==3.0.0
EditorConfig==0.12.3
html-tag-names==0.1.2
html-void-elements==0.1.0
jsbeautifier==1.14.7
json5==0.9.14
pathspec==0.11.1
PyYAML==6.0
regex==2023.5.5
six==1.16.0
sqlparse==0.4.4
tqdm==4.65.0
tzdata==2023.3
